rearrange cd code, add AUTO_PUSHD
